Always clip to *greater* positions when wrapAtSoftNewlines is true
This adjusts the behavior of DisplayBuffer::clipScreenPosition to clip
to a greater position if wrapAtSoftNewlines is true. This requires
`wrapAtSoftNewlines: true` to be passed when moving the cursor down,
which makes intuitive sense.

I also adjusted many specs to account for correctly clipping backward
by one column when moving backward across a soft-wrap boundary.